backend/requirements.txt
fastapi>=0.100.0
uvicorn[standard]>=0.23.0
sqlalchemy>=2.0.0
pydantic>=2.0.0
httpx>=0.25.0
beautifulsoup4>=4.12.0
anthropic>=0.25.0
python-dotenv>=1.0.0
lxml>=4.9.0
duckduckgo-search>=6.0.0
